Anda el normalizador

+++ TP_Metamapa/ModuloAVD/src/main/java/Repositorio/UbicacionRepositorio.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UbicacionRepositorio extends JpaRepository<Ubicacion, Integer> {
+public interface UbicacionRepositorio extends JpaRepository<Ubicacion, Long> {
     Ubicacion findByLocalidadAndProvinciaAndPais(Localidad localidad, Provincia provincia, Pais pais);
 }
